Mega Refactoring

+++ src/main/java/com/harium/etyl/core/animation/Animator.java
@@ -1,5 +1,5 @@
 package com.harium.etyl.core.animation;
 
 public interface Animator {
-    void animate(double factor);
+    void animate(float factor);
 }
