+++ src/main/java/com/harium/etyl/core/animation/script/RotateAnimation.java
@@ -22,7 +22,7 @@ public class RotateAnimation extends SingleIntervalAnimation {
     }
 
     @Override
-    public void update(double value) {
+    public void update(float value) {
         target.setAngle(value);
     }
 }
